installed some flow-typed libdefs and corrected the tests which were failing following the typing in the previous commits

+++ src/lib/duck/reducers/gameReducer/players/playerOptionsToState/optionsConverter/index.js
@@ -7,7 +7,7 @@ export const optionsConverter = (playerOptions: PlayerOptions): GamePlayer => {
   const gamePlayer = {
     start_coord_x: parseInt(playerOptions.start_coord_x),
     start_coord_y: parseInt(playerOptions.start_coord_y),
-    direction: parseInt(playerOptions.direction),
+    direction: parseFloat(playerOptions.direction),
     turn_left_keycode: parseInt(playerOptions.turn_left_keycode),
     turn_right_keycode: parseInt(playerOptions.turn_right_keycode),
     colour: playerOptions.colour
